// web-app/src/types/admin.types.ts
export interface Employee {
  uid: string
  code: string
  name: string
  loginId: string
  email?: string
  phone?: string
  status: 'Active' | 'Inactive'
  authType: string
  createdBy?: string
  createdDate?: string
  modifiedBy?: string
  modifiedDate?: string
  serverModifiedTime?: string
  actionType?: 'Add' | 'Update' | 'Delete'
}